+++ crates/ewwii/src/server.rs
@@ -87,10 +87,7 @@ pub fn initialize_server<B: DisplayBackend>(
         instance_id_to_args: HashMap::new(),
         css_provider: gtk4::CssProvider::new(),
         custom_css_providers: Vec::new(),
-        plugin_buffer: {
-            let (tx, _rx) = tokio::sync::broadcast::channel(64);
-            plugin::PluginBuffer { tx, _rx }
-        },
+        plugin_buffer: plugin::PluginBuffer::new(),
         reloading: false,
         app_evt_send: ui_send.clone(),
         window_close_timer_abort_senders: HashMap::new(),
@@ -117,9 +114,7 @@ pub fn initialize_server<B: DisplayBackend>(
     match read_config {
         Ok(new_config) => {
             app.ewwii_config = new_config;
-            if let Err(e) = app.plugin_buffer.tx.send("ewwii-config-loaded".to_string()) {
-                log::error!("Ewwii failed to emit signal: {e}");
-            }
+            app.plugin_buffer.emit("ewwii-config-loaded");
         }
         Err(err) => {
             error_handling_ctx::print_error(err);
@@ -134,9 +129,7 @@ pub fn initialize_server<B: DisplayBackend>(
         for css_provider in &app.custom_css_providers {
             gtk4::style_context_add_provider_for_display(&display, css_provider, 910);
         }
-        if let Err(e) = app.plugin_buffer.tx.send("ewwii-applied-styles".to_string()) {
-            log::error!("Ewwii failed to emit signal: {e}");
-        }
+        app.plugin_buffer.emit("ewwii-applied-styles");
     }
 
     if let Ok((file_id, css)) = config::scss::parse_scss_from_config(app.paths.get_config_dir()) {